Ecs and godot35

+++ plugins/formats/image/jpg/image_loader_jpegd.cpp
@@ -108,7 +108,7 @@ Error jpeg_load_image_from_buffer(ImageData &p_image, const uint8_t *p_buffer, i
 Error ImageLoaderJPG::load_image(ImageData &p_image, FileAccess *f, LoadParams params) {
 
     PoolVector<uint8_t> src_image;
-    int src_image_len = f->get_len();
+    uint64_t src_image_len = f->get_len();
     ERR_FAIL_COND_V(src_image_len == 0, ERR_FILE_CORRUPT);
     src_image.resize(src_image_len);
